Added More methods to EmployeeController

+++ src/main/java/com/mysqlwebservices/mysqlws/controllers/employeeController.java
@@ -3,6 +3,7 @@ package com.mysqlwebservices.mysqlws.controllers;
 import com.mysqlwebservices.mysqlws.dao.employeeDAO;
 import com.mysqlwebservices.mysqlws.model.Employee;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -25,14 +26,44 @@ public class employeeController {
     }
 
     //get all employees
-    @GetMapping
+    @GetMapping("/employees")
     public List<Employee> getEmployees(){
         return employeeDAO.allEmployees();
     }
 
-    @DeleteMapping
-    public void deleteEmployee(Employee emp){
+    //get employee by id
+    @GetMapping("/employees/{id}")
+    public ResponseEntity<Employee> getEmployeeById(@PathVariable(value = "id") Long empid){
+        Employee emp = employeeDAO.employeeById(empid);
+        if (emp == null){
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok().body(emp);
+    }
+
+    //update an employee by id
+    @PutMapping("/employees/{id}")
+    public ResponseEntity<Employee> updateEmployee(@PathVariable(value = "id") Long empid, @Valid @RequestBody Employee empDetails){
+        Employee emp = employeeDAO.employeeById(empid);
+        if (emp == null){
+            return ResponseEntity.notFound().build();
+        }
+        emp.setName(empDetails.getName());
+        emp.setDesignation(empDetails.getDesignation());
+        emp.setExpertise(empDetails.getExpertise());
+        //check to see if we don't have to do a save after updating this and
+        //Employee updatedEmployee = employeeDAO.newEmployee(emp);
+        //return ResponseEntity.ok().body(updatedEmployee);
+        return ResponseEntity.ok().body(emp);
+    }
+    @DeleteMapping("/employees/{id}")
+    public ResponseEntity<Employee> deleteEmployee(@PathVariable(value = "id") Long empid){
+        Employee emp = employeeDAO.employeeById(empid);
+        if (emp == null){
+            return ResponseEntity.notFound().build();
+        }
         employeeDAO.deleteEmployee(emp);
+        return ResponseEntity.ok().build();
     }
 
 }
